#71 ゲーム画面へstate情報の変更を反映できるようにしました
